requirements.txt
Flask
pytesseract
pillow
pymupdf
python-docx
openpyxl
python-pptx